// dashboard-frontend/src/styles/theme/types.d.ts
export type ColorMode = 'dark' | 'light';